+++ superset-frontend/src/DodoExtensions/onBoarding/index.ts
@@ -1,2 +1,4 @@
 export * from './onBoardingEntryPoint';
 export * from './utils/onboardingMenuItems';
+export { onboardingTeamSlice } from './model/slice/teamSlice';
+export { onboardingSlice } from './model/slice/onboardingSlice';
