+++ superset-frontend/src/DodoExtensions/onBoarding/model/selector/getTeamsData.ts
@@ -1,6 +1,8 @@
-import { OnboardingState } from '../types';
+import { OnboardingTeamState } from '../types/team.types';
 
-export const getTeamsData = (state: { onboarding: OnboardingState }) => ({
-  teamsIsLoading: state.onboarding.teamsIsLoading,
-  teams: state.onboarding.teams,
+export const getTeamsData = (state: {
+  onboardingTeam: OnboardingTeamState;
+}) => ({
+  teamsIsLoading: state.onboardingTeam.teamsIsLoading,
+  teams: state.onboardingTeam.teams,
 });
