+++ superset-frontend/src/DodoExtensions/onBoarding/model/slice/teamSlice.ts
@@ -1,0 +1,38 @@
+import {
+  ONBOARDING_TEAMS_ERROR,
+  ONBOARDING_TEAMS_LOADING,
+  ONBOARDING_TEAMS_SUCCESS,
+  OnboardingTeamAction,
+  OnboardingTeamState,
+} from '../types/team.types';
+
+const initialState: OnboardingTeamState = {
+  teamsIsLoading: false,
+  teams: [],
+  teamsLoadingError: null,
+};
+
+export const onboardingTeamSlice = (
+  state: OnboardingTeamState = initialState,
+  action: OnboardingTeamAction,
+): OnboardingTeamState => {
+  switch (action.type) {
+    case ONBOARDING_TEAMS_LOADING: {
+      return { ...state, teamsIsLoading: true };
+    }
+    case ONBOARDING_TEAMS_SUCCESS: {
+      return { ...state, teamsIsLoading: false, teams: action.payload };
+    }
+    case ONBOARDING_TEAMS_ERROR: {
+      return {
+        ...state,
+        teamsIsLoading: false,
+        teamsLoadingError: action.payload.error,
+      };
+    }
+
+    default: {
+      return state;
+    }
+  }
+};
