+++ superset-frontend/src/views/store.ts
@@ -49,7 +49,10 @@ import { AnyDatasourcesAction } from 'src/explore/actions/datasourcesActions';
 import { HydrateExplore } from 'src/explore/actions/hydrateExplore';
 import getBootstrapData from 'src/utils/getBootstrapData';
 import { Dataset } from '@superset-ui/chart-controls';
-import { onboardingSlice } from '../DodoExtensions/onBoarding/model/slice/onboardingSlice';
+import {
+  onboardingSlice,
+  onboardingTeamSlice,
+} from '../DodoExtensions/onBoarding';
 
 // Some slice don't do anything, and redux is just used to reference the initial "state".
 // This may change later, as the client application takes on more responsibilities.
@@ -130,7 +133,9 @@ const reducers = {
   reports,
   saveModal,
   explore,
+  // DODO added
   onboarding: onboardingSlice,
+  onboardingTeam: onboardingTeamSlice,
 };
 
 /* In some cases the jinja template injects two seperate React apps into basic.html
